requirements.txt
alembic==1.7.5
cffi==1.15.0
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
Flask==2.0.2
Flask-Caching==1.10.1
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gevent==21.12.0
greenlet==1.1.2p
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.0
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
redis==4.0.2
SQLAlchemy==1.4.28
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zope.event==4.5.0
zope.interface==5.4.0
